Try new structure
Test deterministic migration
